+++ RedPandaIDE/settings.h
@@ -355,8 +355,14 @@ public:
         bool enableLigaturesSupport() const;
         void setEnableLigaturesSupport(bool newEnableLigaturesSupport);
 
-        const QString &nonAsciiFontName() const;
-        void setNonAsciiFontName(const QString &newNonAsciiFontName);
+        const QString &fallbackFontName() const;
+        void setFallbackFontName(const QString &newFontName);
+
+        const QString &fallbackFontName2() const;
+        void setFallbackFontName2(const QString &newFontName);
+
+        const QString &fallbackFontName3() const;
+        void setFallbackFontName3(const QString &newFontName);
 
         int mouseSelectionScrollSpeed() const;
         void setMouseSelectionScrollSpeed(int newMouseSelectionScrollSpeed);
@@ -447,7 +453,9 @@ public:
         //Font
         //font
         QString mFontName;
-        QString mNonAsciiFontName;
+        QString mFallbackFontName;
+        QString mFallbackFontName2;
+        QString mFallbackFontName3;
         int mFontSize;
         bool mFontOnlyMonospaced;
         double mLineSpacing;