+++ RedPandaIDE/translations/RedPandaIDE_pt_BR.ts
@@ -1483,7 +1483,7 @@
     </message>
     <message>
         <source>Font for non-ascii Text:</source>
-        <translation>Fonte para texto não ASCII:</translation>
+        <translation type="vanished">Fonte para texto não ASCII:</translation>
     </message>
     <message>
         <source>Gutter</source>
@@ -1551,6 +1551,22 @@
     </message>
     <message>
         <source>Line break</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Fallback Font:</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>*Needs restart</source>
+        <translation type="unfinished">*Necessário reiniciar</translation>
+    </message>
+    <message>
+        <source>Fallback Font 2:</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Fallback Font 3:</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
@@ -5416,11 +5432,11 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>Line: %1 Char: %2 Sel:%3 Lines: %4</source>
+        <source>Line: %1/%2 Char: %3/%4 Sel:%5</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>Line: %1 Char: %2 Lines: %3</source>
+        <source>Line: %1/%2 Char: %3/%4</source>
         <translation type="unfinished"></translation>
     </message>
 </context>